分时图 k线

+++ app/src/main/java/com/yjjr/yjfutures/utils/StringUtils.java
@@ -255,7 +255,7 @@ public class StringUtils {
     }
 
 
-    public static void setOnlineTxArrow(TextView tv, int change) {
+    public static void setOnlineTxArrow(TextView tv, double change) {
         if (change > 0) {
             tv.setText("↑");
         } else if (change < 0) {
